add forked tests
